services/api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.1"
description = "API for Critino, the LLM Critique Platform."
authors = [
    "Jonas Lindberg <jonas.lindberg@futi.no>",
    "Jorge Lewis <jorge.lewis@futi.no>",
]
readme = "./README.md"
homepage = "https://critino.starti.no"
repository = "https://github.com/startino/critino"
documentation = "https://api.critino.starti.no/docs"
license = "BSL-1.1 (https://github.com/startino/critino/blob/alpha/LICENSE)"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = { extras = ["standard"], version = "*" }
tenacity = "*"
supabase = "*"
langchain = "*"
langchain-community = "*"
langchain-openai = "*"
langchain_google_genai = "*"
youtube_transcript_api = "*"
praw = "*"
markdown = "*"
gptrim = "*"
selenium = "*"
python-dotenv = "*"
diskcache = "*"
pydantic = "*"
sse-starlette = "*"
watchfiles = "0.23.0"
sentence-transformers = "^3.3.1"

[tool.poetry.group.dev.dependencies]
mypy = "*"
black = "*"
isort = "*"
autoflake = "*"
flake8 = "*"
pytest = "*"
pytest-cov = "*"
# pyright = "*"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
